Move test framework into merge
