+++ merge/update_test.go
@@ -14,9 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package merge_test
 
 import (
+	"testing"
+
 	"sigs.k8s.io/structured-merge-diff/merge"
 	"sigs.k8s.io/structured-merge-diff/typed"
 )
@@ -57,4 +59,57 @@ func (s *State) Apply(obj typed.TypedValue, owner string, force bool) error {
 	s.Owners = owners
 
 	return nil
+}
+
+// TestExample shows how to use the test framework
+func TestExample(t *testing.T) {
+	state := &State{Updater: &merge.Updater{}}
+	parser, err := typed.NewParser(`types:
+- name: lists
+  struct:
+    fields:
+    - name: list
+      type:
+        list:
+          elementType:
+            scalar: string
+          elementRelationship: associative`)
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+
+	config, err := parser.FromYAML(`
+list:
+- a
+- b
+- c
+`, "lists")
+	if err != nil {
+		t.Fatalf("Failed to parse YAML: %v", err)
+	}
+	err = state.Apply(config, "default", false)
+	if err != nil {
+		t.Fatalf("Wanted err = %v, got %v", nil, err)
+	}
+
+	config, err = parser.FromYAML(`
+list:
+- a
+- b
+- c
+- d`, "lists")
+	if err != nil {
+		t.Fatalf("Failed to parse YAML: %v", err)
+	}
+	err = state.Apply(config, "default", false)
+
+	if err != nil {
+		t.Fatalf("Wanted err = %v, got %v", nil, err)
+	}
+
+	// The following is wrong because the code doesn't work yet.
+	_, err = state.Live.Compare(config)
+	if err == nil {
+		t.Fatalf("Succeeded to compare live with config")
+	}
 }
